moving everything for the reviewModal to a new folder to help with organization
