+++ client/src/components/Review Widget/AddReview/AddReviewModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect }  from 'react';
-import './cssFiles/reviewSection.css';
+import '../cssFiles/reviewSection.css';
 import axios from 'axios'
 import CharacteristicsBar from './CharacteristicsBar.jsx';
 import AddReviewStars from './AddReviewStars.jsx';
@@ -111,6 +111,11 @@ const AddReviewModal = (props) => {
     let formRating = addReviewHelpers.getFormStarRating();
     let formRecommended = addReviewHelpers.getFormRecommended();
     let formCharacteristics = addReviewHelpers.getFormCharacteristics();
+    let formReviewSummary = addReviewHelpers.getFormReviewSummary();
+    let formReviewBody = addReviewHelpers.getFormReviewBody();
+    let formReviewNickName = addReviewHelpers.getFormReviewNickName();
+    let formReviewEmail = addReviewHelpers.getFormReviewEmail();
+
     console.log('forms');
 
 
@@ -169,13 +174,13 @@ const AddReviewModal = (props) => {
             <div>
               <label style={addReviewTitleStyle}>4. Review Summary:</label><br></br>
               <div style={addReviewItemStyle}>
-                <input type="textarea" maxLength="60" placeholder="Example: Best purhcase ever!" style={{width: '90%', height: '30px', fontFamily: 'Playfair Display serif'}}/>
+                <input id="reviewSummary" type="textarea" maxLength="60" placeholder="Example: Best purhcase ever!" style={{width: '90%', height: '30px', fontFamily: 'Playfair Display serif'}}/>
               </div>
             </div>
             <div>
               <label style={addReviewTitleStyle}>5. Review Body: *</label><br></br>
               <div style={addReviewItemStyle}>
-                <textarea onChange={onReviewBodyChange} maxLength="1000" minLength="50" placeholder="Why did you like the product or not?" style={{width: '90%', height: '100px', fontFamily: 'Playfair Display serif'}}></textarea>
+                <textarea id="reviewBody" onChange={onReviewBodyChange} maxLength="1000" minLength="50" placeholder="Why did you like the product or not?" style={{width: '90%', height: '100px', fontFamily: 'Playfair Display serif'}}></textarea>
                 <br></br>
                 <span style={subTextStyle}>{getCharCountDisplay()}</span>
               </div>
@@ -183,7 +188,7 @@ const AddReviewModal = (props) => {
             <div>
               <label style={addReviewTitleStyle}>6. What is your nickname?: *</label><br></br>
               <div style={addReviewItemStyle}>
-                <input type="text" maxLength="60" placeholder="Example: jackson11!" style={{width: '70%', fontFamily: 'Playfair Display serif'}}/>
+                <input id="nickName" type="text" maxLength="60" placeholder="Example: jackson11!" style={{width: '70%', fontFamily: 'Playfair Display serif'}}/>
                 <br></br>
                 <span style={subTextStyle}>{'For privacy reasons, do not use your full name or email address'}</span>
               </div>
@@ -191,7 +196,7 @@ const AddReviewModal = (props) => {
             <div>
               <label style={addReviewTitleStyle}>7. Your email: *</label><br></br>
               <div style={addReviewItemStyle}>
-                <input type="text" maxLength="60" style={{width: '70%', fontFamily: 'Playfair Display serif'}} placeholder="Example: jackson11@email.com"/>
+                <input id="email" type="text" maxLength="60" style={{width: '70%', fontFamily: 'Playfair Display serif'}} placeholder="Example: jackson11@email.com"/>
                 <br></br>
                 <span style={subTextStyle}>{'For authentication reasons, you will not be not be emailed'}</span>
               </div>
